requirements.txt
fastapi
uvicorn
jinja2
requests
python-dotenv
pillow

📦 Install dependencies:
pip install -r requirements.txt
